\name{addFile}
\alias{addFile}
\title{Stores a file route in a ini file, for loading when arqas starts.}
\usage{
  addFile(file)
}
\arguments{
  \item{file}{Route to the file}
}
\value{
  TRUE if everything goes well
}
\description{
  Stores a file route in a ini file, for loading when arqas
  starts.
}

