% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arqasgui.R
\docType{data}
\name{exportedFunctions}
\alias{exportedFunctions}
\title{Counter of exported funtions}
\format{\preformatted{ num 20
}}
\usage{
exportedFunctions
}
\description{
Counter of exported funtions
}
\keyword{internal}

