% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arqasgui.R
\name{modelarg}
\alias{modelarg}
\title{Create a list with the label and the type of the argument of the model}
\usage{
modelarg(label, type)
}
\arguments{
\item{label}{Description of the argument}

\item{type}{Type of the argument (numeric, boolean, vector, matrix, distr, vdistr)}
}
\description{
Create a list with the label and the type of the argument of the model
}

