% Generated by roxygen2 (4.0.2): do not edit by hand
\name{registerDistribution}
\alias{registerDistribution}
\title{Register a distribution to the UI}
\usage{
registerDistribution(fun, name, index = FALSE)
}
\arguments{
\item{fun}{Function of the distribution}

\item{name}{Name of the distribution}

\item{index}{Create a label in the list with the name of the distribution}
}
\description{
Register a distribution to the UI
}

