% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{registeredDistributions}
\alias{registeredDistributions}
\title{Counter of registered distributions}
\format{\preformatted{ num 9
}}
\usage{
registeredDistributions
}
\description{
Counter of registered distributions
}
\keyword{internal}

