% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arqasgui.R
\docType{data}
\name{uiList}
\alias{uiList}
\title{List of export models to the UI}
\format{\preformatted{List of 20
 $ :List of 4
  ..$ id  : num 1
  ..$ name: chr "M/M/1"
  ..$ fun :function (lambda = 3, mu = 6)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 26 10 45 1 10 1 380 399
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x000000000d2de8b8> 
  ..$ args:List of 2
  .. ..$ :List of 2
  .. .. ..$ label: chr "Arrival rate"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Service rate"
  .. .. ..$ type : chr "numeric"
  ..- attr(*, "class")= chr [1:3] "M_M_1" "M_M_S" "MarkovianModel"
 $ :List of 4
  ..$ id  : num 2
  ..$ name: chr "M/M/s"
  ..$ fun :function (lambda = 3, mu = 6, s = 2)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 110 10 139 1 10 1 464 493
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x000000000d2de8b8> 
  ..$ args:List of 3
  .. ..$ :List of 2
  .. .. ..$ label: chr "Arrival rate"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Service rate"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Servers"
  .. .. ..$ type : chr "numeric"
  ..- attr(*, "class")= chr [1:2] "M_M_S" "MarkovianModel"
 $ :List of 4
  ..$ id  : num 3
  ..$ name: chr "M/M/1/K"
  ..$ fun :function (lambda = 3, mu = 6, k = 2)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 220 12 249 1 12 1 574 603
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x000000000d2de8b8> 
  ..$ args:List of 3
  .. ..$ :List of 2
  .. .. ..$ label: chr "Arrival rate"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Service rate"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Queue size"
  .. .. ..$ type : chr "numeric"
  ..- attr(*, "class")= chr [1:3] "M_M_1_K" "M_M_S_K" "MarkovianModel"
 $ :List of 4
  ..$ id  : num 4
  ..$ name: chr "M/M/s/K"
  ..$ fun :function (lambda = 3, mu = 6, s = 2, k = 3)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 360 12 401 1 12 1 714 755
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x000000000d2de8b8> 
  ..$ args:List of 4
  .. ..$ :List of 2
  .. .. ..$ label: chr "Arrival rate"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Service rate"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Servers"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Queue size"
  .. .. ..$ type : chr "numeric"
  ..- attr(*, "class")= chr [1:2] "M_M_S_K" "MarkovianModel"
 $ :List of 4
  ..$ id  : num 5
  ..$ name: chr "M/M/1/INF/H"
  ..$ fun :function (lambda = 1/2, mu = 60/5, h = 5)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 37 16 66 1 16 1 891 920
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x000000000d290648> 
  ..$ args:List of 3
  .. ..$ :List of 2
  .. .. ..$ label: chr "Arrival rate"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Service rate"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Potential population"
  .. .. ..$ type : chr "numeric"
  ..- attr(*, "class")= chr [1:3] "M_M_1_INF_H" "M_M_S_INF_H" "MarkovianModel"
 $ :List of 4
  ..$ id  : num 6
  ..$ name: chr "M/M/s/INF/H"
  ..$ fun :function (lambda = 1/2, mu = 60/5, s = 2, h = 5)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 157 16 187 1 16 1 1011 1041
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x000000000d290648> 
  ..$ args:List of 4
  .. ..$ :List of 2
  .. .. ..$ label: chr "Arrival rate"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Service rate"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Servers"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Potential population"
  .. .. ..$ type : chr "numeric"
  ..- attr(*, "class")= chr [1:2] "M_M_S_INF_H" "MarkovianModel"
 $ :List of 4
  ..$ id  : num 7
  ..$ name: chr "M/M/s/INF/H with Y replacements"
  ..$ fun :function (lambda = 3, mu = 6, s = 3, h = 5, y = 3)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 307 18 349 1 18 1 1161 1203
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x000000000d290648> 
  ..$ args:List of 5
  .. ..$ :List of 2
  .. .. ..$ label: chr "Arrival rate"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Service rate"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Servers"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Potential population"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Replacements"
  .. .. ..$ type : chr "numeric"
  ..- attr(*, "class")= chr [1:3] "M_M_S_INF_H_Y" "M_M_S_INF_H" "MarkovianModel"
 $ :List of 4
  ..$ id  : num 8
  ..$ name: chr "M/M/INF"
  ..$ fun :function (lambda = 3, mu = 6)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 442 12 458 1 12 1 1296 1312
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x000000000d290648> 
  ..$ args:List of 2
  .. ..$ :List of 2
  .. .. ..$ label: chr "Arrival rate"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Service rate"
  .. .. ..$ type : chr "numeric"
  ..- attr(*, "class")= chr [1:2] "M_M_INF" "MarkovianModel"
 $ :List of 4
  ..$ id  : num 9
  ..$ name: chr "Open Jackson Network"
  ..$ fun :function (lambda = c(20, 30), mu = c(100, 25), s = c(1, 2), p = matrix(c(0.2, 
    0.25, 0.1, 0), nrow = 2, ncol = 2))  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 57 23 106 1 23 1 1435 1484
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x000000000d243d30> 
  ..$ args:List of 4
  .. ..$ :List of 2
  .. .. ..$ label: chr "Arrival rate at each node"
  .. .. ..$ type : chr "vector"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Service rate at each node"
  .. .. ..$ type : chr "vector"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Servers at each node"
  .. .. ..$ type : chr "vector"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Transition matrix"
  .. .. ..$ type : chr "matrix"
  ..- attr(*, "class")= chr [1:3] "OpenJackson" "Network" "MarkovianModel"
 $ :List of 4
  ..$ id  : num 10
  ..$ name: chr "Closed Jackson Network"
  ..$ fun :function (mu = c(5, 5, 10, 15), s = c(2, 2, 1, 1), p = matrix(c(0.25, 0.15, 
    0.2, 0.4, 0.15, 0.35, 0.2, 0.3, 0.5, 0.25, 0.15, 0.1, 0.4, 0.3, 0.25, 
    0.05), 4, byrow = TRUE), n = 3)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 280 25 355 1 25 1 1658 1733
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x000000000d243d30> 
  ..$ args:List of 4
  .. ..$ :List of 2
  .. .. ..$ label: chr "Service rate at each node"
  .. .. ..$ type : chr "vector"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Servers at each node"
  .. .. ..$ type : chr "vector"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Transition matrix"
  .. .. ..$ type : chr "matrix"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Number of customers"
  .. .. ..$ type : chr "numeric"
  ..- attr(*, "class")= chr [1:3] "ClosedJackson" "Network" "MarkovianModel"
 $ :List of 4
  ..$ id  : num 11
  ..$ name: chr "G/G/1"
  ..$ fun :function (arrivalDistribution = Exp(3), serviceDistribution = Exp(6), staClients = 100, 
    nClients = 1000, historic = FALSE, nsim = 10, nproc = 1)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 153 10 293 1 10 1 1945 2085
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x000000000d1f4c68> 
  ..$ args:List of 7
  .. ..$ :List of 2
  .. .. ..$ label: chr "Arrival distribution"
  .. .. ..$ type : chr "distr"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Service distribution"
  .. .. ..$ type : chr "distr"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Stabilization parameter"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Simulation parameter"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Generate historic?"
  .. .. ..$ type : chr "boolean"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Number of simulations"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Number of subprocesses"
  .. .. ..$ type : chr "numeric"
  ..- attr(*, "class")= chr [1:2] "G_G_1" "SimulatedModel"
 $ :List of 4
  ..$ id  : num 12
  ..$ name: chr "G/G/s"
  ..$ fun :function (arrivalDistribution = Exp(3), serviceDistribution = Exp(6), s = 2, 
    staClients = 100, nClients = 1000, historic = FALSE, nsim = 10, nproc = 1)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 322 10 492 1 10 1 2114 2284
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x000000000d1f4c68> 
  ..$ args:List of 8
  .. ..$ :List of 2
  .. .. ..$ label: chr "Arrival distribution"
  .. .. ..$ type : chr "distr"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Service distribution"
  .. .. ..$ type : chr "distr"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Servers"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Stabilization parameter"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Simulation parameter"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Generate historic?"
  .. .. ..$ type : chr "boolean"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Number of simulations"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Number of subprocesses"
  .. .. ..$ type : chr "numeric"
  ..- attr(*, "class")= chr [1:2] "G_G_S" "SimulatedModel"
 $ :List of 4
  ..$ id  : num 13
  ..$ name: chr "G/G/1/K"
  ..$ fun :function (arrivalDistribution = Exp(3), serviceDistribution = Exp(6), K = 2, 
    staClients = 100, nClients = 1000, historic = FALSE, nsim = 10, nproc = 1)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 521 12 668 1 12 1 2313 2460
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x000000000d1f4c68> 
  ..$ args:List of 8
  .. ..$ :List of 2
  .. .. ..$ label: chr "Arrival distribution"
  .. .. ..$ type : chr "distr"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Service distribution"
  .. .. ..$ type : chr "distr"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Queue size"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Stabilization parameter"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Simulation parameter"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Generate historic?"
  .. .. ..$ type : chr "boolean"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Number of simulations"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Number of subprocesses"
  .. .. ..$ type : chr "numeric"
  ..- attr(*, "class")= chr [1:2] "G_G_1_K" "SimulatedModel"
 $ :List of 4
  ..$ id  : num 14
  ..$ name: chr "G/G/s/K"
  ..$ fun :function (arrivalDistribution = Exp(3), serviceDistribution = Exp(6), s = 2, 
    K = 3, staClients = 100, nClients = 1000, historic = FALSE, nsim = 10, 
    nproc = 1)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 698 12 862 1 12 1 2490 2654
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x000000000d1f4c68> 
  ..$ args:List of 9
  .. ..$ :List of 2
  .. .. ..$ label: chr "Arrival distribution"
  .. .. ..$ type : chr "distr"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Service distribution"
  .. .. ..$ type : chr "distr"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Servers"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Queue size"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Stabilization parameter"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Simulation parameter"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Generate historic?"
  .. .. ..$ type : chr "boolean"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Number of simulations"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Number of subprocesses"
  .. .. ..$ type : chr "numeric"
  ..- attr(*, "class")= chr [1:2] "G_G_S_K" "SimulatedModel"
 $ :List of 4
  ..$ id  : num 15
  ..$ name: chr "G/G/1/INF/H"
  ..$ fun :function (arrivalDistribution = Exp(3), serviceDistribution = Exp(6), H = 5, 
    staClients = 100, nClients = 1000, historic = FALSE, nsim = 10, nproc = 1)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 892 16 1034 1 16 1 2684 2826
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x000000000d1f4c68> 
  ..$ args:List of 8
  .. ..$ :List of 2
  .. .. ..$ label: chr "Arrival distribution"
  .. .. ..$ type : chr "distr"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Service distribution"
  .. .. ..$ type : chr "distr"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Potential customers"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Stabilization parameter"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Simulation parameter"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Generate historic?"
  .. .. ..$ type : chr "boolean"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Number of simulations"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Number of subprocesses"
  .. .. ..$ type : chr "numeric"
  ..- attr(*, "class")= chr [1:2] "G_G_1_INF_H" "SimulatedModel"
 $ :List of 4
  ..$ id  : num 16
  ..$ name: chr "G/G/s/INF/H"
  ..$ fun :function (arrivalDistribution = Exp(3), serviceDistribution = Exp(6), s = 3, 
    H = 5, staClients = 100, nClients = 1000, historic = FALSE, nsim = 10, 
    nproc = 1)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 1064 16 1235 1 16 1 2856 3027
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x000000000d1f4c68> 
  ..$ args:List of 9
  .. ..$ :List of 2
  .. .. ..$ label: chr "Arrival distribution"
  .. .. ..$ type : chr "distr"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Service distribution"
  .. .. ..$ type : chr "distr"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Servers"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Potential customers"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Stabilization parameter"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Simulation parameter"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Generate historic?"
  .. .. ..$ type : chr "boolean"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Number of simulations"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Number of subprocesses"
  .. .. ..$ type : chr "numeric"
  ..- attr(*, "class")= chr [1:2] "G_G_S_INF_H" "SimulatedModel"
 $ :List of 4
  ..$ id  : num 17
  ..$ name: chr "G/G/s/INF/H with Y replacements"
  ..$ fun :function (arrivalDistribution = Exp(3), serviceDistribution = Exp(6), s = 3, 
    H = 5, Y = 3, staClients = 100, nClients = 1000, historic = FALSE, 
    nsim = 10, nproc = 1)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 1265 18 1457 1 18 1 3057 3249
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x000000000d1f4c68> 
  ..$ args:List of 10
  .. ..$ :List of 2
  .. .. ..$ label: chr "Arrival distribution"
  .. .. ..$ type : chr "distr"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Service distribution"
  .. .. ..$ type : chr "distr"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Servers"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Potential customers"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Replacements"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Stabilization parameter"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Simulation parameter"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Generate historic?"
  .. .. ..$ type : chr "boolean"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Number of simulations"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Number of subprocesses"
  .. .. ..$ type : chr "numeric"
  ..- attr(*, "class")= chr [1:2] "G_G_S_INF_H_Y" "SimulatedModel"
 $ :List of 4
  ..$ id  : num 18
  ..$ name: chr "G/G/INF"
  ..$ fun :function (arrivalDistribution = Exp(3), serviceDistribution = Exp(6), staClients = 100, 
    nClients = 1000, historic = FALSE, nsim = 10, nproc = 1)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 1484 12 1619 1 12 1 3276 3411
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x000000000d1f4c68> 
  ..$ args:List of 7
  .. ..$ :List of 2
  .. .. ..$ label: chr "Arrival distribution"
  .. .. ..$ type : chr "distr"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Service distribution"
  .. .. ..$ type : chr "distr"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Stabilization parameter"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Simulation parameter"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Generate historic?"
  .. .. ..$ type : chr "boolean"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Number of simulations"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Number of subprocesses"
  .. .. ..$ type : chr "numeric"
  ..- attr(*, "class")= chr [1:2] "G_G_INF" "SimulatedModel"
 $ :List of 4
  ..$ id  : num 19
  ..$ name: chr "Open Network"
  ..$ fun :function (arrivalDistribution = c(Exp(20), Exp(30)), serviceDistribution = c(Exp(100), 
    Exp(25)), s = c(1, 2), p = matrix(c(0.2, 0.25, 0.1, 0), nrow = 2, ncol = 2), 
    staClients = 100, transitions = 1000, historic = FALSE, nsim = 10, 
    nproc = 1)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 1900 16 2188 1 16 1 3692 3980
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x000000000d1f4c68> 
  ..$ args:List of 9
  .. ..$ :List of 2
  .. .. ..$ label: chr "Arrival distribution at each node"
  .. .. ..$ type : chr "vdistr"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Service distribution at each node"
  .. .. ..$ type : chr "vdistr"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Servers at each node"
  .. .. ..$ type : chr "vector"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Transition matrix"
  .. .. ..$ type : chr "matrix"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Stabilization parameter"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Simulation parameter"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Generate historic?"
  .. .. ..$ type : chr "boolean"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Number of simulations"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Number of subprocesses"
  .. .. ..$ type : chr "numeric"
  ..- attr(*, "class")= chr [1:3] "Open" "SimulatedNetwork" "SimulatedModel"
 $ :List of 4
  ..$ id  : num 20
  ..$ name: chr "Closed Network"
  ..$ fun :function (serviceDistribution = c(Exp(5), Exp(5), Exp(10), Exp(15)), s = c(2, 
    2, 1, 1), p = array(c(0.25, 0.15, 0.5, 0.4, 0.15, 0.35, 0.25, 0.3, 
    0.2, 0.2, 0.15, 0.25, 0.4, 0.3, 0.1, 0.05), dim = c(4, 4)), staClients = 100, 
    nClients = 3, transitions = 1000, historic = FALSE, nsim = 10, nproc = 1)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 1666 18 1854 1 18 1 3458 3646
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x000000000d1f4c68> 
  ..$ args:List of 9
  .. ..$ :List of 2
  .. .. ..$ label: chr "Service distribution at each node"
  .. .. ..$ type : chr "vdistr"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Servers at each node"
  .. .. ..$ type : chr "vector"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Transition matrix"
  .. .. ..$ type : chr "matrix"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Number of customers"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Stabilization parameter"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Simulation parameter"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Generate historic?"
  .. .. ..$ type : chr "boolean"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Number of simulations"
  .. .. ..$ type : chr "numeric"
  .. ..$ :List of 2
  .. .. ..$ label: chr "Number of subprocesses"
  .. .. ..$ type : chr "numeric"
  ..- attr(*, "class")= chr [1:3] "Closed" "SimulatedNetwork" "SimulatedModel"
}}
\usage{
uiList
}
\description{
List of export models to the UI
}
\keyword{internal}

