% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arqasgui.R
\docType{data}
\name{distrList}
\alias{distrList}
\title{List of registered distributions to the UI}
\format{\preformatted{List of 2
 $ get:function ()  
 $ set:function (v)  
}}
\usage{
distrList
}
\description{
List of registered distributions to the UI
}
\keyword{internal}

