\docType{data}
\name{distrList}
\alias{distrList}
\title{List of registered distributions to the UI}
\format{List of 8
 $ Exp    :List of 3
 $ Norm   :List of 3
 $ Weibull:List of 3
 $ Unif   :List of 3
 $ Chisq  :List of 3
 $ Lnorm  :List of 3
 $ Gammad :List of 3
 $ Dirac  :List of 3}
\usage{
  distrList
}
\description{
  List of registered distributions to the UI
}
\keyword{datasets}
\keyword{internal}

