\docType{data}
\name{distrList}
\alias{distrList}
\title{List of registered distributions to the UI}
\format{List of 6
 $ :List of 3
 $ :List of 3
 $ :List of 3
 $ :List of 3
 $ :List of 3
 $ :List of 3}
\usage{
  distrList
}
\description{
  List of registered distributions to the UI
}
\keyword{datasets}
\keyword{internal}

