% Generated by roxygen2 (4.0.2): do not edit by hand
\name{exportToUI}
\alias{exportToUI}
\title{Exports a function to the UI}
\usage{
exportToUI(fun, name, args, class)
}
\arguments{
\item{fun}{Function of the model}

\item{name}{Name of the model}

\item{args}{Type of each parameter of the function (numerical, character, vector, matrix)}

\item{class}{A string to agrupate funtions in the same menu option}
}
\description{
Exports a function to the UI
}

