\docType{data}
\name{uiList}
\alias{uiList}
\title{List of export models to the UI}
\format{List of 20
 $ :List of 4
  ..- attr(*, "class")= chr [1:3] "M_M_1" "M_M_S" "MarkovianModel"
 $ :List of 4
  ..- attr(*, "class")= chr [1:2] "M_M_S" "MarkovianModel"
 $ :List of 4
  ..- attr(*, "class")= chr [1:3] "M_M_1_K" "M_M_S_K" "MarkovianModel"
 $ :List of 4
  ..- attr(*, "class")= chr [1:2] "M_M_S_K" "MarkovianModel"
 $ :List of 4
  ..- attr(*, "class")= chr [1:3] "M_M_1_INF_H" "M_M_S_INF_H" "MarkovianModel"
 $ :List of 4
  ..- attr(*, "class")= chr [1:2] "M_M_S_INF_H" "MarkovianModel"
 $ :List of 4
  ..- attr(*, "class")= chr [1:3] "M_M_S_INF_H_Y" "M_M_S_INF_H" "MarkovianModel"
 $ :List of 4
  ..- attr(*, "class")= chr [1:2] "M_M_INF" "MarkovianModel"
 $ :List of 4
  ..- attr(*, "class")= chr [1:3] "OpenJackson" "Network" "MarkovianModel"
 $ :List of 4
  ..- attr(*, "class")= chr [1:3] "ClosedJackson" "Network" "MarkovianModel"
 $ :List of 4
  ..- attr(*, "class")= chr [1:2] "G_G_1" "SimulatedModel"
 $ :List of 4
  ..- attr(*, "class")= chr [1:2] "G_G_S" "SimulatedModel"
 $ :List of 4
  ..- attr(*, "class")= chr [1:2] "G_G_1_K" "SimulatedModel"
 $ :List of 4
  ..- attr(*, "class")= chr [1:2] "G_G_S_K" "SimulatedModel"
 $ :List of 4
  ..- attr(*, "class")= chr [1:2] "G_G_1_INF_H" "SimulatedModel"
 $ :List of 4
  ..- attr(*, "class")= chr [1:2] "G_G_S_INF_H" "SimulatedModel"
 $ :List of 4
  ..- attr(*, "class")= chr [1:2] "G_G_S_INF_H_Y" "SimulatedModel"
 $ :List of 4
  ..- attr(*, "class")= chr [1:2] "G_G_INF" "SimulatedModel"
 $ :List of 4
  ..- attr(*, "class")= chr [1:3] "Closed" "SimulatedNetwork" "SimulatedModel"
 $ :List of 4
  ..- attr(*, "class")= chr [1:3] "Open" "SimulatedNetwork" "SimulatedModel"}
\usage{
  uiList
}
\description{
  List of export models to the UI
}
\keyword{datasets}
\keyword{internal}

